Add constructor for SerialTransport and make the necessary modifications in CmdMessenger.h and CmdMessenger.cpp to work with the new abstraction
in CmdMessenger.h, the constrcutor was modified to include the new param
TransportLayer* transport, to attach the transport layer. Alsom, the
private member serial_port_ was modified to transport_layer_ to
represent the transport layer abstraction methodology.

In SerialTransport.h the constructor was updated to include the
necessary params to initialize a Serial object correctly, the return
type of the public method read was changed so that it may return the
number of bytes actually read.

In TransportLayer.h the header v8stdint.h was included to use default
definitions of ints, for a more cross-platform implementation. The read
method had its return type modified to return the number of bytes read
and it was included the method that receives a string as argument.

In CmdMessenger.cpp all ocurrences of serial_port_ was replaced by
transport_layer_.